backend/chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.1.0
langchain-openai==0.0.6
chromadb==0.4.18
requests==2.31.0
python-multipart==0.0.6
